refactor(tests): consolidate OmniVoice and HiggsAudio tests
Move tests from test_omnivoice.py and test_higgs_audio.py into the
shared test_models.py. Add encode pipeline tests, sanitize encode-path
tests, and parity tests with real weights.
